postgres_ffi: rename record to walrecord
