+++ libs/postgres_ffi/src/lib.rs
@@ -216,8 +216,8 @@ macro_rules! enum_pgversion {
 }
 
 pub mod pg_constants;
-pub mod record;
 pub mod relfile_utils;
+pub mod walrecord;
 
 // Export some widely used datatypes that are unlikely to change across Postgres versions
 pub use v14::bindings::RepOriginId;
